ISSUE #3738 - move select logic to assignees component which can be disabled to remove select
